#  Small runtime optimizations (variable lookup speeds)